fix option return for iterator functions
fix returns concrete type
remove concrete and interfaces package for a better usability
