--- concrete/utils.go
@@ -1,5 +1,0 @@
-package concrete
-
-func compare(a interface{}, b interface{}) bool {
-	return a == b
-}
