+++ constraints.go
@@ -1,4 +1,4 @@
-package interfaces
+package iterago
 
 type Unsigned interface {
 	uint | uint8 | uint16 | uint32 | uint64 | uintptr
