+++ errors.go
@@ -1,4 +1,4 @@
-package interfaces
+package iterago
 
 const (
 	ErrUnwrapNoneOption = "cannot unwrap a none option"
