+++ go.mod
@@ -2,10 +2,10 @@ module github.com/ulphidius/iterago
 
 go 1.19
 
+require github.com/stretchr/testify v1.8.1
+
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/stretchr/objx v0.5.0 // indirect
-	github.com/stretchr/testify v1.8.1 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
