+++ interface.go
@@ -1,0 +1,6 @@
+package iterago
+
+type Iterator[T any] interface {
+	Next() Option[T]
+	HasNext() bool
+}
