+++ utils_test.go
@@ -1,4 +1,4 @@
-package interfaces
+package iterago
 
 import (
 	"testing"
